Lists of combinations are treated properly now. The map example showcases it by applying Had to a list of booleans.
- Still need to work on properly discarding the values with zero
amplitude. There are some concerns with program inversion and the
representation of classical isos. Right now every clause presents a
combination, with classical isos going 1v1' +0v2' + ... and so forth.
Inversion and typechecking take that into consideration, as do the
application of a linear combination to a classical iso. From those, the
presence of vals with amplitude 0 can mess with the typechecking of
isos, so we need to take care when constructing the clauses involving
variables.
Take the stupid example:
  () <-> 1~() + 0~y
  x  <-> let y = f x in 0~() + 1~y
Typechecking will fail because y is not present in the context of the
first clause definition.